OPENDNSSEC-636: Attempt to fix the build on Jenkins build machines.  This is tested to work on Ubuntu 14.04.1, previously failing.
See discussion on:
http://stackoverflow.com/questions/17055279/autotools-for-pthreads-not-setting-correct-linker-flags
http://stackoverflow.com/questions/17055279/autotools-for-pthreads-not-setting-correct-linker-flags
https://lists.gnu.org/archive/html/autoconf-archive-maintainers/2011-11/msg00001.html
http://lists.gnu.org/archive/html/autoconf-archive-maintainers/2013-06/msg00021.html
This seems to be a problem with the AX_PTHREADS macro, which is not repaired
yet in version #21 of the m4 pthread macro (I have upgraded the macro
nevertheless).  The problem occurs when C++ files are introduced, where
the PTHREAD_CFLAGS are not included.
There is no proper fix, but the fix by adding the flags in the AX_PTHREAD
is least intrusive.