OPENDNSSEC-302, OPENDNSSEC-652: - realization and usage of incremental time leap in test framework.
Using ods_time_leap_over you can indicate how much time must minimal progress
(in seconds).  It will then progress for that much time.  With a second
parameter you can give a maximum number of time leaps should be taken, as
a kind-of timeout parameter.  If the maximum is reached, it is considered
a failure.