OPENDNSSEC-637: The -p option to the enforcer daemon is no longer there.  This option would be used to perform an enforce run on a single policy. With the current setup, where the enforce task always does all policies, this is not there anymore.  Although not impossible, it is probably better to split up the enforce task before (if ever) restoring this option.
I think the main feature is for testing purpose, as it requires you
to run the enforcer in one-shot modus constantly.
It has only been available since 1.4.3 and there is only a single line
of documentation about it.  I cannot find a proper reason for it on
mailing list or wiki, and I imagine that this has been developer pushed.

Dropping this test.