OPENDNSSEC-726, OPENDNSSEC-731:
Initially tried to add just a single section to the signconf file,
but after looking at how the KSK public key and the signature of
the signed records get stored, it is more usedfull to do this
separately.

It turned out the dnskey record of the KSK must be filled at some
time otherwise many parts of the code will fail.