OPENDNSSEC-726, OPENDNSSEC-735, OPENDNSSEC-739: Tool to bundle signed keyset and public key of KSK into signconf. This is more of a make-shift tool for the moment, usable for the unit test.
It was expected that the signed zone would contain enough information
to integrate the DNSKEY set into the signconf.  However, because it
was too hard to use just a single entry containing the signed dnskey
and the public key material of the KSK, there are now multiple sections.
This was due to the signer code always expecting a dnskey record
in the individual key structures.  Splitting this out requires the
right key in the signed zone to be matched against the signconf.
However the signed zone only contains a keytag, which isn't really
unique.  The signconf contains locators, which are unique.

It is also easier to use the backup file after shutdown to retrieve
the key record from, because there the locator and the keytag can
be matched.