Make schedule lock recursive
The reason we allow recursive locks is because we broadcast
a condition in an ISR. The main thread would receive this
interrupt. If we would do no locking in the ISR we risk missing
conditions. If we use a normal mutex we risk a deadlock if the
main thread occasionally grabs the lock.