Revert "Merge pull request #508 from yschaeff/drifting_time"
This reverts commit c72e9171e07da06f49a9f535ade9dd5a6a924a44, reversing
changes made to 7109da28ea4bfc9f16fd7676b5f747eed0ec8d19.

The changes in themselves are suppostedly correct, but they break the tests and
the tests probably need revising.  As this however causes other problems to
go unnoticed (because the build is broken).  I'm reverting the commit
to get into a stable state again.