SUPPORT-265: The issue where "key_data_update" was shown in the log proved to be worse.  There isn't a conflict between two operations but within the same operation in the enforcer.  Whilst enforcing a zone and making a step in the key roll, when also doing a key removal from the hsm this will occur.
The main problem is within the database layer not being able
to manage this, but this cannot be resolved at this time.
Also the deleting of keys has nothing to do with the key roll
so should be done as a separate task.  There is no such task
at this time.

By switching around operations we can avoid the conflict,
but this does have the drawback that the keys won't be
deleted immediately but on the next run.