When listing keys avoid calling C_FindObjects for every key
The existing implementation finds all private keys and then looks for the
public keys with the same ID. If the HSM has an inefficient implementation
of C_FindObjects for the public key lookup this makes the list process
slow.

To avoid doing multiple calls to C_FindObjects, find all of the public keys
and their IDs then pair them together. This results in double the number of
calls to C_GetAttributeValue (to get the public key IDs) but as this is
already used to get the ID for every private key it can't make the listing time
take significantly longer than before.