backup_hsmkeys_cmd: fix option handling
`longgetopt()` reorders `argv` so that non-options are at the end, i.e.
changing `{"backup", "list", "-r", "SoftHSMv2"}` to `{"-r", "SoftHSMv2",
"backup", "list"}`. This means that later uses of `argv[1]` in the function
will compare the wrong string with the possible subcommands like `list`. This
effectively broke the `-r` option completely.

This commit fixes this by not passing the first two values from `argv` to
`longgetopt()` as these contain "backup" and the subcommand.