OPENDNSSEC-964: No keys exported when exporting all keys with mysql back-end
When a query is submitted without parameters, because there are non using this
particular call, it is not necessary to verify the existence of a parameter list.
Within the subsequent while loop, when trying to fill parameters, this check
is done already.  Which is the proper place as at that time there indeed should
be parameters.