Fix longgetopt usage
`longgetopt()` is a functional compatible version of `getopt_long()`.
As illustrated in the [man page of `getopt_long()`](https://man.freebsd.org/cgi/man.cgi?getopt_long(3)), after parsing the options, the `argc` may be substracted with `optind` and `argv` may be incremented by `optind` to continue processing non-option arguments.
Because `longgetopt()` is a thread safe version not using global variables, the `optind` is in the `optctx` struct that is passed along calls to `longgetopt()`