Fix portability of `struct option` and friends
`struct option` is a companion of the `getopt_long()` function, and is repurposed for the functional equivalent `longgetopt()` function.
An integral part of the `struct option` definition are the possible values of its `has_arg` struct member.
If `struct option` needs to be defined, because it was detected during configure that `getopt_long()` is not available, then also the valuees of the `has_arg` struct member need to be defined.

Please see [the man page for `getopt_long()`](https://man.freebsd.org/cgi/man.cgi?getopt_long(3)) to confirm that `no_argument`, `required_argument` and `optional_argument` are required for proper usage of `struct option` with `getopt_long()`, and because `longgetopt()` is functionally equivalent with `getopt_long()`, also with our custom `longgetopt()`